// libs/util-words/src/lib/models/word-category.model.ts
export enum WordCategory {
  Animals = 'Animals',
  Sports = 'Sports',
  Countries = 'Countries',
  Objects = 'Objects',
  Colors = 'Colors',
  Transportation = 'Transportation',
  PublicPlaces = 'PublicPlaces',
  Nature = 'Nature',
  Kitchen = 'Kitchen',
  Beauty = 'Beauty',
  MoviesAndSeries = 'MoviesAndSeries',
  // Food = 'Food',
  // 'Games' = 'Games',
  // 'Cities' = 'Cities',
  // 'Brands' = 'Brands',
  // 'Celebrities' = 'Celebrities',
  // 'Professions' = 'Professions',
}